+++ src/anomalib/models/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2022-2025 Intel Corporation
+# Copyright (C) 2022-2026 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 """Anomaly detection models.
@@ -151,23 +151,28 @@ def list_models(case: str = "snake") -> set[str]:
         >>> # Get models in snake_case format
         >>> models = list_models(case="snake")
         >>> print(sorted(list(models)))  # doctest: +NORMALIZE_WHITESPACE
-        ['ai_vad', 'cfa', 'cflow', 'csflow', 'dfkde', 'dfm', 'draem',
-         'efficient_ad', 'fastflow', 'fre', 'ganomaly', 'padim', 'patchcore',
-         'reverse_distillation', 'stfpm', 'uflow', 'vlm_ad', 'winclip']
+        ['ai_vad', 'anomaly_d_i_n_o', 'cfa', 'cflow', 'csflow', 'dfkde', 'dfm',
+         'dinomaly', 'draem', 'dsr', 'efficient_ad', 'fastflow', 'fre', 'fuvas',
+         'ganomaly', 'general_a_d', 'padim', 'patchcore',
+         'reverse_distillation', 'stfpm', 'supersimplenet', 'uflow', 'uni_net',
+         'vlm_ad', 'win_clip']
 
         >>> # Get models in original PascalCase format
         >>> models = list_models(case="pascal")
         >>> print(sorted(list(models)))  # doctest: +NORMALIZE_WHITESPACE
-        ['AiVad', 'Cfa', 'Cflow', 'Csflow', 'Dfkde', 'Dfm', 'Draem',
-         'EfficientAd', 'Fastflow', 'Fre', 'Ganomaly', 'Padim', 'Patchcore',
-         'ReverseDistillation', 'Stfpm', 'Uflow', 'VlmAd', 'WinClip']
+        ['AiVad', 'AnomalyDINO', 'Cfa', 'Cflow', 'Csflow', 'Dfkde', 'Dfm',
+         'Dinomaly', 'Draem', 'Dsr', 'EfficientAd', 'Fastflow', 'Fre', 'Fuvas',
+         'Ganomaly', 'GeneralAD', 'Padim', 'Patchcore', 'ReverseDistillation',
+         'Stfpm', 'Supersimplenet', 'Uflow', 'UniNet', 'VlmAd', 'WinClip']
 
         >>> # Get models in title case format
         >>> models = list_models(case="title")
         >>> print(sorted(list(models)))  # doctest: +NORMALIZE_WHITESPACE
-        ['Ai Vad', 'Cfa', 'Cflow', 'Csflow', 'Dfkde', 'Dfm', 'Draem',
-         'Efficient Ad', 'Fastflow', 'Fre', 'Ganomaly', 'Padim', 'Patchcore',
-         'Reverse Distillation', 'Stfpm', 'Uflow', 'Vlm Ad', 'Win Clip']
+        ['Ai Vad', 'Anomaly Dino', 'Cfa', 'Cflow', 'Csflow', 'Dfkde', 'Dfm',
+         'Dinomaly', 'Draem', 'Dsr', 'Efficient Ad', 'Fastflow', 'Fre', 'Fuvas',
+         'Ganomaly', 'General Ad', 'Padim', 'Patchcore',
+         'Reverse Distillation', 'Stfpm', 'Supersimplenet', 'Uflow', 'Uni Net',
+         'Vlm Ad', 'Win Clip']
 
     Note:
         The returned model names can be used with :func:`get_model` to instantiate